Refactor ProspectStatus resource structure: move pages and resource to ProspectStatuses namespace, remove old files, and update tests accordingly.
